Fix exception import issue and move exceptions.
