+++ examples/rllib/rllib_agent.py
@@ -11,7 +11,7 @@ try:
     from ray.rllib.models.tf.fcnet import FullyConnectedNetwork
     from ray.rllib.utils import try_import_tf
 except Exception as e:
-    from .. import RayException
+    from smarts.core.utils.custom_exceptions import RayException
 
     raise RayException.required_to("rllib_agent.py")
