+++ examples/tools/regression_rllib.py
@@ -14,7 +14,7 @@ try:
     from ray.rllib.models import ModelCatalog
     from ray.rllib.utils import try_import_tf
 except Exception as e:
-    from .. import RayException
+    from smarts.core.utils.custom_exceptions import RayException
 
     raise RayException.required_to("regression_rllib.py")
