+++ examples/tools/stress_sumo.py
@@ -1,7 +1,7 @@
 try:
     import ray
 except Exception as e:
-    from .. import RayException
+    from smarts.core.utils.custom_exceptions import RayException
 
     raise RayException.required_to("stress_sumo.py")
